revert some changes
